+++ src/main/java/com/example/Homework_230809/services/interfaces/CalculatorService.java
@@ -2,8 +2,8 @@ package com.example.Homework_230809.services.interfaces;
 
 public interface CalculatorService {
     String getGreetingMessage();
-    Integer add(Integer num1, Integer num2);
-    Integer subtract(Integer num1, Integer num2);
-    Integer multiply(Integer num1, Integer num2);
-    Integer divide(Integer num1, Integer num2);
+    Long add(Integer num1, Integer num2);
+    Long subtract(Integer num1, Integer num2);
+    Long multiply(Integer num1, Integer num2);
+    Double divide(Integer num1, Integer num2);
 }
